+++ tests/testthat/_problems/test-pipeline-gss-125.R
@@ -1,5 +1,0 @@
-# Extracted from test-pipeline-gss.R:125
-
-# 
-# Failed to extract test: 
-# Failed to find test at line 125.