+++ tests/testthat/_problems/test-pipeline-gss-148.R
@@ -1,5 +1,0 @@
-# Extracted from test-pipeline-gss.R:148
-
-# 
-# Failed to extract test: 
-# Failed to find test at line 148.